vastly speed up text rendering
Turns out h-raylib's drawTextEx allocates on the heap and marshals for
each call:

- the Font
- the Vector2
- the Color
- the CString

This turns out to be extremely slow, meaning 2 FPS for a small page of
text on my machine. Passing around the raw pointers and getting rid of
Vector2 in the call altogether increases this more than a hundred times.